Reorganize sym.rs into bit.rs and vec.rs
