+++ crates/sym/src/lib.rs
@@ -1,15 +1,18 @@
+mod bit;
 mod buf;
 mod convert;
 mod eval;
 mod pcode;
-mod sym;
+mod vec;
 
+pub use crate::bit::*;
 pub use crate::buf::*;
 pub use crate::convert::ConcreteValue;
+pub use crate::convert::ConcretizationError;
 pub use crate::convert::concretize;
 pub use crate::convert::concretize_into;
 pub use crate::eval::*;
-pub use crate::sym::*;
+pub use crate::vec::*;
 
 #[cfg(test)]
 mod tests;
